Functions to add/remove Layers/Sources

+++ Naxam.Mapbox.Forms/Sources/RasterSource.cs
@@ -4,7 +4,7 @@
     {
         public const int DEFAULT_TILE_SIZE = 512;
 
-        public RasterSource(string id, string configurationURL, double tileSize = DEFAULT_TILE_SIZE)
+        public RasterSource(string id, string configurationURL, int tileSize = DEFAULT_TILE_SIZE)
         {
             Id = id;
             ConfigurationURL = configurationURL;
@@ -13,6 +13,6 @@
 
         public string ConfigurationURL { get; private set; }
 
-        public double TileSize { get; private set; }
+        public int TileSize { get; private set; }
     }
 }
